+++ CHANGELOG.md
@@ -3,6 +3,14 @@ Added this change log after project was already under way.  Early changes are no
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [Unreleased]
+### Changed
+- Added JavaDoc to `ListArray`
+
+### Added
+- New class `ListMap`
+- New class `SortedListMap`
+
 ## [1.2] - 2017-02-19
 ### Added
 - New class `ReaderBuffer`